Use new enumeration query
